Add comparePassword instance method to User model
